update text and button colors feature added
