+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 0.0.1
-
 * initial Release
+
+## 0.0.2
+* we can update colors & style of button and text
