+++ example/integration_test/plugin_integration_test.dart
@@ -6,11 +6,8 @@
 // For more information about Flutter integration tests, please see
 // https://docs.flutter.dev/cookbook/testing/integration/introduction
 
-
 import 'package:integration_test/integration_test.dart';
-
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
-
 }
