+++ example/lib/image_resource.dart
@@ -1,6 +1,5 @@
 class ImageResouce {
-  static const android2= "assets/images/android_2.png";
-  static const android3= "assets/images/android_3.png";
-  static const android4= "assets/images/android_4.png";
+  static const android2 = "assets/images/android_2.png";
+  static const android3 = "assets/images/android_3.png";
+  static const android4 = "assets/images/android_4.png";
 }
-
