--- example/lib/second_screen.dart
@@ -1,18 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SecondScreen extends StatelessWidget {
-  const SecondScreen({super.key});
-  static const routeName = "/secondScreen";
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-        alignment: Alignment.center,
-        child: const Text("second Screen"),
-      ),
-    );
-  }
-}
-
-
