+++ example/test/second_screen_test.dart
@@ -1,13 +1,12 @@
-import 'package:app_walk_through_example/second_screen.dart';
+import 'package:app_walk_through_example/main.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'my_widget_tester.dart';
 
-void main(){
+void main() {
   testWidgets("should have AppWalkThrough widget", (widgetTester) async {
     await widgetTester.pumpWidget(const MyWidgetTester(screen: SecondScreen()));
     final textFinder = find.text("second Screen");
     expect(textFinder, findsOneWidget);
   });
 }
-
