+++ lib/src/app_walk_through_data_model.dart
@@ -1,7 +1,11 @@
+/// this is appWalkThrough data model which will be helpful to define app short descriptions.
+/// assign appWalkThrough item [title] text,
+/// assign appWalkThrough item [description] text &
+/// assign assets [image] string of appWalkThrough item
+
 class AppWalkThroughDataModel {
   String? title;
   String? description;
   String? image;
   AppWalkThroughDataModel({this.title, this.description, this.image});
 }
-
